requirements.txt
pandas==1.2.4
line_bot_sdk==1.19.0
Flask==1.1.2
requests >= 2.0